+++ src/models/casos.ts
